chore(release): version packages (#121)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @folks-finance/xchain-sdk@0.0.15

### Patch Changes

-
[#120](https://github.com/Folks-Finance/folks-finance-xchain-js-sdk/pull/120)
[`0dbde90`](https://github.com/Folks-Finance/folks-finance-xchain-js-sdk/commit/0dbde906fd5b3755bae37b29120aca9c9c9ee793)
Thanks [@gidonkatten](https://github.com/gidonkatten)! - deployment of
bsc and bnb

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>